Move PPM/PFM to distinct module
