+++ Texconv/Texconv_Desktop_2019.vcxproj.filters
@@ -8,6 +8,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Texconv.cpp" />
+    <ClCompile Include="PortablePixMap.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="Texconv.rc">
